# Added Shape Algorithms and Optimized Line Drawing