… TheSadRogue.Primitives.UnitTests/LinesTests.cs
… TheSadRogue.Primitives.UnitTests/ShapesTests.cs
… TheSadRogue.Primitives/Lines.cs
… TheSadRogue.Primitives/ShapeAlgorithms.cs
… TheSadRogue.Primitives/Shapes.cs