# Changes
- Added algorithms to draw circles and ellipses
- Modified all line and shape algorithms to use custom iterators, which makes them significantly faster
- Added line-algorithm specific functions (for example, `GetBresehnamLine`, `GetDDALine`) to the `Lines` class
    - Users should now use these instead of `GetLine` unless they need an `IEnumerable` specifically or they need a single function to which you can specify any supported line algorithm, as they are significantly faster

# Notes
This PR uses the same custom-iterator struct strategy used by other portions of the library (`Rectangle.Positions`, for example) in order to create algorithms that allow users to iterate over their results significantly faster.  the line algorithms are about 2x faster now, and the circle algorithms are around the same compared to other implementations.

One side effect of this change, is that each line algorithm has a separate return type.  The return types can't really be exposed as a single interface, since doing so would require boxing the structures and that would have many of the same performance characteristics `IEnumerable` does.

As is typically the case with custom iterators in the primitives library, you _can_ still get an `IEnumerable<T>` by calling `ToEnumerable()` on the iterator.  As such, there is a generic `GetLine(Point, Point, Lines.Algorithm)` function, which uses this functionality to return an `IEnumerable` which iterates over the line as created by the specified drawing algorithm (this is equivalent to the original interface).  However, if a user needs optimal performance, unless they _really_ need the ability to specify the algorithm to a single function, they're likely better off just using the algorithm-specific function, since it will perform significantly better.

Additionally, the PR also adds algorithms for drawing circles and ellipses.  They similarly use custom iterators, which have a `ToEnumerable()` function in case a user actually needs an `IEnumerable`.

## Validation of Correctness
Custom iterators, although very fast, are rather challenging to write correctly for complex algorithms.  Some of the code is, at best, an eyesore.  Furthermore, `ToEnumerable` is often a re-implementation of the algorithm entirely, rather than a use of the iterator in a foreach loop, to avoid additional overhead, making the code rather difficult to manage.

In an effort to validate correctness, the unit tests do the following things:
- Test the results of `ToEnumerable()` and the faster custom iterator against each other; they must be the same for all test cases
- Test the results of the custom iterator against a simpler, more direct implementation of the algorithm, which isn't fast but is much easier to write correctly.  The results must match (plus or minus order)
- Test the results against typical constraints expected (for example, lines must have adjacent points, be within the bounds of start and end, etc.

This, combined with code coverage metrics, should provide a relatively robust framework for ensuring the algorithms are correct.

# Benchmarks
In the below benchmarks:
- **Primitives:** Uses the current primitives library implementation by using the custom iterator directly in a foreach loop
- **PrimitivesToEnumerable:** Uses the current primitives library implementation, but calls `ToEnumerable()` and iterates over that
- **Func:** An implementation which returns nothing but takes a `Func<Point, bool>` as a parameter instead, and calls that on every point (similar to SadConsole's original implementation
- **YieldReturn:** A simple implementation which simply returns each point by using `yield return`.

## Line Algorithms
|                           Method |       LineToDraw |      Mean |    Error |   StdDev |
|--------------------------------- |----------------- |----------:|---------:|---------:|
|  BresenhamPrimitivesToEnumerable |  ((1,1), (1,25)) | 203.51 ns | 0.724 ns | 0.642 ns |
|                    BresenhamFunc |  ((1,1), (1,25)) | 178.24 ns | 0.251 ns | 0.210 ns |
|              BresenhamPrimitives |  ((1,1), (1,25)) |  81.80 ns | 0.549 ns | 0.487 ns |
|             BresenhamYieldReturn |  ((1,1), (1,25)) | 188.14 ns | 0.551 ns | 0.460 ns |
|        DDAPrimitivesToEnumerable |  ((1,1), (1,25)) | 185.03 ns | 3.240 ns | 2.872 ns |
|                    DDAPrimitives |  ((1,1), (1,25)) |  82.15 ns | 0.497 ns | 0.388 ns |
|                   DDAYieldReturn |  ((1,1), (1,25)) | 185.24 ns | 0.701 ns | 0.655 ns |
| OrthogonalPrimitivesToEnumerable |  ((1,1), (1,25)) | 202.84 ns | 0.680 ns | 0.531 ns |
|             OrthogonalPrimitives |  ((1,1), (1,25)) |  74.57 ns | 0.295 ns | 0.262 ns |
|            OrthogonalYieldReturn |  ((1,1), (1,25)) | 195.33 ns | 1.792 ns | 1.677 ns |
|  BresenhamPrimitivesToEnumerable | ((1,1), (25,50)) | 369.78 ns | 1.182 ns | 0.923 ns |
|                    BresenhamFunc | ((1,1), (25,50)) | 348.07 ns | 0.516 ns | 0.403 ns |
|              BresenhamPrimitives | ((1,1), (25,50)) | 146.34 ns | 0.232 ns | 0.181 ns |
|             BresenhamYieldReturn | ((1,1), (25,50)) | 365.17 ns | 0.453 ns | 0.378 ns |
|        DDAPrimitivesToEnumerable | ((1,1), (25,50)) | 369.94 ns | 0.288 ns | 0.225 ns |
|                    DDAPrimitives | ((1,1), (25,50)) | 156.41 ns | 0.694 ns | 0.580 ns |
|                   DDAYieldReturn | ((1,1), (25,50)) | 373.64 ns | 1.019 ns | 0.796 ns |
| OrthogonalPrimitivesToEnumerable | ((1,1), (25,50)) | 556.98 ns | 5.793 ns | 5.135 ns |
|             OrthogonalPrimitives | ((1,1), (25,50)) | 199.19 ns | 1.297 ns | 1.214 ns |
|            OrthogonalYieldReturn | ((1,1), (25,50)) | 534.52 ns | 1.978 ns | 1.651 ns |
|  BresenhamPrimitivesToEnumerable |  ((1,1), (50,1)) | 366.78 ns | 1.860 ns | 1.649 ns |
|                    BresenhamFunc |  ((1,1), (50,1)) | 338.35 ns | 1.597 ns | 1.494 ns |
|              BresenhamPrimitives |  ((1,1), (50,1)) | 137.87 ns | 0.661 ns | 0.586 ns |
|             BresenhamYieldReturn |  ((1,1), (50,1)) | 354.47 ns | 1.891 ns | 1.677 ns |
|        DDAPrimitivesToEnumerable |  ((1,1), (50,1)) | 341.07 ns | 3.264 ns | 3.053 ns |
|                    DDAPrimitives |  ((1,1), (50,1)) | 140.22 ns | 0.688 ns | 0.644 ns |
|                   DDAYieldReturn |  ((1,1), (50,1)) | 355.79 ns | 1.622 ns | 1.354 ns |
| OrthogonalPrimitivesToEnumerable |  ((1,1), (50,1)) | 378.43 ns | 1.469 ns | 1.374 ns |
|             OrthogonalPrimitives |  ((1,1), (50,1)) | 133.83 ns | 0.764 ns | 0.715 ns |
|            OrthogonalYieldReturn |  ((1,1), (50,1)) | 364.75 ns | 2.021 ns | 1.891 ns |
|  BresenhamPrimitivesToEnumerable | ((1,1), (50,25)) | 368.48 ns | 1.630 ns | 1.525 ns |
|                    BresenhamFunc | ((1,1), (50,25)) | 344.46 ns | 1.548 ns | 1.448 ns |
|              BresenhamPrimitives | ((1,1), (50,25)) | 144.33 ns | 0.763 ns | 0.676 ns |
|             BresenhamYieldReturn | ((1,1), (50,25)) | 359.38 ns | 1.397 ns | 1.239 ns |
|        DDAPrimitivesToEnumerable | ((1,1), (50,25)) | 364.21 ns | 1.576 ns | 1.397 ns |
|                    DDAPrimitives | ((1,1), (50,25)) | 154.75 ns | 1.015 ns | 0.848 ns |
|                   DDAYieldReturn | ((1,1), (50,25)) | 369.77 ns | 1.243 ns | 1.038 ns |
| OrthogonalPrimitivesToEnumerable | ((1,1), (50,25)) | 547.81 ns | 3.266 ns | 3.055 ns |
|             OrthogonalPrimitives | ((1,1), (50,25)) | 206.43 ns | 0.783 ns | 0.733 ns |
|            OrthogonalYieldReturn | ((1,1), (50,25)) | 534.12 ns | 2.110 ns | 1.871 ns |
|  BresenhamPrimitivesToEnumerable | ((25,50), (1,1)) | 371.78 ns | 3.033 ns | 2.689 ns |
|                    BresenhamFunc | ((25,50), (1,1)) | 345.17 ns | 2.049 ns | 1.917 ns |
|              BresenhamPrimitives | ((25,50), (1,1)) | 145.26 ns | 0.636 ns | 0.563 ns |
|             BresenhamYieldReturn | ((25,50), (1,1)) | 360.20 ns | 1.629 ns | 1.524 ns |
|        DDAPrimitivesToEnumerable | ((25,50), (1,1)) | 364.75 ns | 1.989 ns | 1.861 ns |
|                    DDAPrimitives | ((25,50), (1,1)) | 154.59 ns | 0.710 ns | 0.593 ns |
|                   DDAYieldReturn | ((25,50), (1,1)) | 383.14 ns | 2.400 ns | 2.245 ns |
| OrthogonalPrimitivesToEnumerable | ((25,50), (1,1)) | 547.54 ns | 4.944 ns | 4.129 ns |
|             OrthogonalPrimitives | ((25,50), (1,1)) | 198.41 ns | 1.516 ns | 1.266 ns |
|            OrthogonalYieldReturn | ((25,50), (1,1)) | 535.93 ns | 2.756 ns | 2.443 ns |
|  BresenhamPrimitivesToEnumerable | ((50,25), (1,1)) | 368.34 ns | 1.910 ns | 1.693 ns |
|                    BresenhamFunc | ((50,25), (1,1)) | 343.48 ns | 1.979 ns | 1.755 ns |
|              BresenhamPrimitives | ((50,25), (1,1)) | 144.88 ns | 0.850 ns | 0.795 ns |
|             BresenhamYieldReturn | ((50,25), (1,1)) | 361.46 ns | 4.834 ns | 4.285 ns |
|        DDAPrimitivesToEnumerable | ((50,25), (1,1)) | 368.23 ns | 0.780 ns | 0.651 ns |
|                    DDAPrimitives | ((50,25), (1,1)) | 156.09 ns | 0.847 ns | 0.707 ns |
|                   DDAYieldReturn | ((50,25), (1,1)) | 375.16 ns | 1.646 ns | 1.459 ns |
| OrthogonalPrimitivesToEnumerable | ((50,25), (1,1)) | 554.03 ns | 0.822 ns | 0.642 ns |
|             OrthogonalPrimitives | ((50,25), (1,1)) | 210.21 ns | 1.575 ns | 1.473 ns |
|            OrthogonalYieldReturn | ((50,25), (1,1)) | 576.39 ns | 6.882 ns | 6.437 ns |

The custom iterator implementations appear to be up to 2x faster.  Calling `ToEnumerable()` on the custom iterator and that is roughly the same as just using an implementation of the algorithm which uses `yield return` (~20ns added for structure creation, I think).

 ## CircleAlgorithm
|                 Method | Radius |        Mean |     Error |    StdDev |
|----------------------- |------- |------------:|----------:|----------:|
|             Primitives |      5 |    86.45 ns |  1.344 ns |  1.191 ns |
| PrimitivesToEnumerable |      5 |   210.86 ns |  2.149 ns |  2.010 ns |
|                   Plot |      5 |   173.14 ns |  0.374 ns |  0.292 ns |
|            YieldReturn |      5 |   196.55 ns |  2.263 ns |  2.006 ns |
|             Primitives |     10 |   171.65 ns |  1.873 ns |  1.564 ns |
| PrimitivesToEnumerable |     10 |   396.68 ns |  3.372 ns |  2.632 ns |
|                   Plot |     10 |   335.13 ns |  1.950 ns |  1.728 ns |
|            YieldReturn |     10 |   386.04 ns |  6.378 ns |  5.966 ns |
|             Primitives |     25 |   466.31 ns |  4.565 ns |  4.270 ns |
| PrimitivesToEnumerable |     25 |   938.03 ns |  2.931 ns |  2.288 ns |
|                   Plot |     25 |   812.67 ns |  0.317 ns |  0.248 ns |
|            YieldReturn |     25 |   919.07 ns |  9.620 ns |  8.033 ns |
|             Primitives |     50 | 1,011.89 ns |  3.479 ns |  2.905 ns |
| PrimitivesToEnumerable |     50 | 1,882.39 ns | 13.924 ns | 12.344 ns |
|                   Plot |     50 | 1,635.32 ns |  1.506 ns |  1.257 ns |
|            YieldReturn |     50 | 1,873.12 ns | 13.878 ns | 12.982 ns |

The custom iterator implementations appear to be a bit under 2x faster.  Calling `ToEnumerable()` on the custom iterator and that is roughly the same as just using an implementation of the algorithm which uses `yield return` (~20ns added for structure creation, I think).

## Ellipse Algorithm
|                 Method |          Ellipse |     Mean |   Error |   StdDev |   Median |
|----------------------- |----------------- |---------:|--------:|---------:|---------:|
|             Primitives |  ((1,1), (1,25)) | 200.1 ns | 6.54 ns | 19.28 ns | 209.2 ns |
| PrimitivesToEnumerable |  ((1,1), (1,25)) | 377.3 ns | 3.33 ns |  2.78 ns | 378.1 ns |
|            YieldReturn |  ((1,1), (1,25)) | 352.3 ns | 6.85 ns |  7.89 ns | 347.9 ns |
|             Primitives | ((1,1), (25,50)) | 383.6 ns | 1.29 ns |  1.07 ns | 383.3 ns |
| PrimitivesToEnumerable | ((1,1), (25,50)) | 777.9 ns | 2.16 ns |  1.69 ns | 777.4 ns |
|            YieldReturn | ((1,1), (25,50)) | 753.6 ns | 2.83 ns |  2.36 ns | 752.5 ns |
|             Primitives |  ((1,1), (50,1)) | 313.9 ns | 1.11 ns |  0.93 ns | 314.1 ns |
| PrimitivesToEnumerable |  ((1,1), (50,1)) | 753.6 ns | 4.96 ns |  4.64 ns | 752.2 ns |
|            YieldReturn |  ((1,1), (50,1)) | 667.5 ns | 1.76 ns |  1.37 ns | 667.2 ns |
|             Primitives | ((25,50), (1,1)) | 406.8 ns | 2.31 ns |  1.93 ns | 406.1 ns |
| PrimitivesToEnumerable | ((25,50), (1,1)) | 782.9 ns | 7.23 ns |  6.76 ns | 779.3 ns |
|            YieldReturn | ((25,50), (1,1)) | 759.7 ns | 1.70 ns |  1.42 ns | 759.5 ns |

The custom iterator implementations appear to be a bit under 2x faster.  Calling `ToEnumerable()` on the custom iterator and that is roughly the same as just using an implementation of the algorithm which uses `yield return` (~20ns added for structure creation, I think).